requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.3.0
idna==3.2
inflection==0.5.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2021.5.29
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
uritemplate==3.0.1
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
